Make defaul import depth configurable for each OPAC
