+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/processFromTemplate/dialogs/import.xhtml
@@ -73,7 +73,7 @@
                         <p:spinner id="importDepth"
                                    class="input"
                                    disabled="#{empty CreateProcessForm.catalogImportDialog.hitModel.selectedCatalog}"
-                                   value="#{CreateProcessForm.catalogImportDialog.importDepth}"
+                                   value="#{CreateProcessForm.catalogImportDialog.hitModel.importDepth}"
                                    min="1"
                                    max="5"/>
                     </div>
